requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
google-genai
python-dotenv
gunicorn
